+++ test/torus/bm25_compile_test.exs
@@ -22,6 +22,37 @@ defmodule Torus.BM25CompileTest do
       end
     end
 
+    test "score_key with a map select before the call compiles" do
+      code = """
+      import Ecto.Query
+      import Torus
+      alias TorusTest.Post
+
+      Post
+      |> select([p], %{body: p.body})
+      |> Torus.bm25([p], p.body, "search", score_key: :relevance)
+      |> limit(5)
+      """
+
+      assert {_query, _bindings} = Code.eval_string(code, [], __ENV__)
+    end
+
+    test "score_key with a select after the call does not compile" do
+      code = """
+      import Ecto.Query
+      import Torus
+      alias TorusTest.Post
+
+      Post
+      |> Torus.bm25([p], p.body, "search", score_key: :relevance)
+      |> select([p], %{body: p.body})
+      """
+
+      assert_raise Ecto.Query.CompileError, ~r/only one select expression is allowed/, fn ->
+        Code.eval_string(code, [], __ENV__)
+      end
+    end
+
     test "score_threshold with index_name compiles successfully" do
       # This should compile without errors
       code = """